lead-ai/crm/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=21.2.0
pydantic>=2.10.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
pandas>=2.2.0
numpy<2.0.0  # Fixed: pandas/pyarrow compatibility
joblib>=1.4.0
python-multipart>=0.0.12
aiohttp>=3.11.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dateutil>=2.8.2
catboost>=1.2
httpx>=0.27.0
jinja2>=3.1.2
twilio>=9.0.0
resend>=2.0.0
sendgrid>=6.11.0
requests>=2.31.0
loguru>=0.7.0
cachetools>=5.3.0
supabase>=2.3.0

# Monitoring & Observability (Optional but recommended for production)
sentry-sdk[fastapi]>=2.0.0  # Error tracking and performance monitoring
prometheus-client>=0.20.0    # Metrics collection and export

# Testing (Optional - for development)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0  # Already included above for async HTTP testing
postgrest>=0.13.0
bcrypt>=4.1.0

# AI Models - Choose one or both
openai>=1.0.0              # For GPT-4 integration
anthropic>=0.18.0          # For Claude integration (Recommended)

email-validator>=2.0.0

# Task Queue & Caching (Phase 3 - Optional)
celery>=5.3.0              # Background task queue
redis>=5.0.0               # Message broker & cache
flower>=2.0.0              # Celery monitoring UI
